cloud_function/requirements.txt
functions-framework>=3.5.0
requests>=2.28.0
google-cloud-storage>=2.0.0
google-auth>=2.20.0
